move the code to delete test database into a command from a test
